+++ run_hot_reload.bat
@@ -1,0 +1,7 @@
+@echo on
+
+cd %~dp0
+call venv\Scripts\activate
+gradio app.py
+
+TIMEOUT /T 10
